renaming and style
