--- src/training/mtool-training-app/requirements.txt
@@ -1,3 +1,0 @@
-keras
-tensorflow
-sklearn
