--- src/training/mtool-training-app/run-spec.yaml
@@ -1,3 +1,0 @@
-name: praxxis-model-training
-version: v1
-inputs: {}
