--- src/training/mtool-training-app/spec.yaml
@@ -1,9 +1,0 @@
-name: praxxis-model-training
-version: v1
-runtime: Python
-src: ./train.py
-entrypoint: train
-replicas: 1
-poolsize: 1
-output:
-  data: str
