--- src/training/mtool-training-app/test.yaml
@@ -1,8 +1,0 @@
-name: praxxis-model-scoring
-version: v1
-inputs:
-   seq:
-      - "SOP023"
-      - "SOP023"
-      - "SOP023"
-
