+++ src/training/praxxis-training-app/requirements.txt
@@ -1,0 +1,3 @@
+keras
+tensorflow
+sklearn
