+++ src/training/praxxis-training-app/run-spec.yaml
@@ -1,0 +1,3 @@
+name: praxxis-training-app
+version: v1
+inputs: {}
