+++ src/training/praxxis-training-app/test.yaml
@@ -1,0 +1,8 @@
+name: praxxis-training-app
+version: v1
+inputs:
+   seq:
+      - "SOP023"
+      - "SOP023"
+      - "SOP023"
+
